Cleanup of options area, implementation of site link system
