+++ dt-core/admin/menu/tab-site-links.php
@@ -1,0 +1,52 @@
+<?php
+
+/**
+ * Disciple Tools
+ *
+ * @class      Disciple_Tools_
+ * @version    0.1.0
+ * @since      0.1.0
+ * @package    Disciple_Tools
+ * @author     Chasm.Solutions & Kingdom.Training
+ */
+
+if ( !defined( 'ABSPATH' ) ) {
+    exit; // Exit if accessed directly
+}
+
+/**
+ * Class Disciple_Tools_Import_Export_Tab
+ */
+class Disciple_Tools_Site_Links_Tab
+{
+    /**
+     * Packages and returns tab page
+     *
+     * @return void
+     */
+    public function content()
+    {
+        ?>
+        <form method="post">
+
+            <div class="wrap">
+                <div id="poststuff">
+                    <div id="post-body" class="metabox-holder columns-2">
+                        <div id="post-body-content">
+
+                            <?php DT_Site_Link_System::metabox_multiple_link(); ?>
+
+                        </div><!-- end post-body-content -->
+                        <div id="postbox-container-1" class="postbox-container">
+                        </div><!-- postbox-container 1 -->
+                        <div id="postbox-container-2" class="postbox-container">
+                        </div><!-- postbox-container 2 -->
+                    </div><!-- post-body meta box container -->
+                </div>
+                <!--poststuff end -->
+            </div><!-- wrap end -->
+        </form>
+        <?php
+    }
+
+}
