+++ functions.php
@@ -407,11 +407,6 @@ else {
                 require_once( get_template_directory() . '/dt-people-groups/admin-menu.php' );
                 $this->people_groups_admin = Disciple_Tools_People_Groups_Admin_Menu::instance();
 
-                // Tables
-                require_once( get_template_directory() . '/dt-core/admin/menu/tables/notifications-table.php' );
-                require_once( get_template_directory() . '/dt-core/admin/menu/tables/activity-list-table.php' ); // contacts and groups report building
-                require_once( get_template_directory() . '/dt-core/admin/menu/tables/reports-list-table.php' ); // contacts and groups report building
-
                 // Metaboxes
                 require_once( get_template_directory() . '/dt-core/admin/metaboxes/box-activity.php' );
                 require_once( get_template_directory() . '/dt-core/admin/metaboxes/box-share-contact.php' );
